serializers for custom types
any of the custom scalar types we want to use, ex: `NonNegativeFloat` will default to `string` unless we create a custom extension and serializer for them.  I've created 4 for the common int and decimal types.  Note I used a `decimal` type for `NonNegativeFloat` even though a `double` might be better, but since this is a financial system I want to start with decimal and go from there.